Update settings.js
fix: the `--temp_dir` argument wasn't being properly passed to Sauce Connect, so would always default to "./temp". Users that did not have a "./temp" directory at the root of their project (and had configured Magellan to point to an alternate path) would get a mysterious exit code 2 from Sauce Connect when it tried to create a log file at the non-existent path. This fixes that issue.